[p2p] bump DEFAULT_MAX_ORPHAN_ANNOUNCEMENTS to 3000
For the default number of peers (125), allows each to relay a default
descendant package of transactions (up to 25-1=24 can be missing inputs)
out of order.

Functional tests aren't changed to check for a cap for 3000
because it would make the runtime too long.

Also deletes the now-unused DEFAULT_MAX_ORPHAN_TRANSACTIONS.